shifted controller to view for now
